backend/requirements.txt
argcomplete==3.6.3
Babel==2.17.0
Beaker==1.13.0
chardet==5.2.0
compiler==0.2.0
ConfigParser==7.2.0
cryptography==46.0.3
docutils==0.22.3
exceptiongroup==1.3.1
fastapi==0.125.0
filelock==3.20.1
HTMLParser==0.0.2
ipython==9.8.0
ipywidgets==8.1.8
jnius==1.1.0
keyring==25.7.0
lingua==4.15.0
numpy==2.3.5
ordereddict==1.1
passlib==1.7.4
pexpect==4.9.0
Pillow==12.0.0
protobuf==6.33.2
pydantic==2.12.5
pyOpenSSL==25.3.0
python-dotenv==1.2.1
python_jose==3.5.0
redis==7.1.0
Sphinx==9.0.4
SQLAlchemy==2.0.45
starlette==0.50.0
thread==2.0.6
Twisted==25.5.0
typing_extensions==4.15.0
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
zope==5.13
